LOGIN & SIGNUP with database
when user sign up and create an account, their credentials are stored in a database and the user and therefore login to the library management system with their respective username and password. Error messages are displayed in cases of error such as if the user does not fill all the blank fields or if username and password does not match.